Move sortResults to sortMultipleViews in DetailsSortUtils
