+++ test/tstDetailsDistributedSearchTreeImpl.cpp
@@ -17,143 +17,7 @@
 
 #define BOOST_TEST_MODULE DetailsDistributedSearchTreeImpl
 
-#include <algorithm> // fill
-#include <set>
-#include <vector>
-
 namespace tt = boost::test_tools;
-
-BOOST_AUTO_TEST_CASE_TEMPLATE(sort_results, DeviceType, ARBORX_DEVICE_TYPES)
-{
-  std::vector<int> ids_ = {4, 3, 2, 1, 4, 3, 2, 4, 3, 4};
-  std::vector<int> sorted_ids = {1, 2, 2, 3, 3, 3, 4, 4, 4, 4};
-  std::vector<int> offset = {0, 1, 3, 6, 10};
-  int const n = 10;
-  int const m = 4;
-  BOOST_TEST(ids_.size() == n);
-  BOOST_TEST(sorted_ids.size() == n);
-  BOOST_TEST(offset.size() == m + 1);
-  std::vector<int> results_ = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
-  std::vector<std::set<int>> sorted_results = {
-      {3},
-      {6, 2},
-      {8, 5, 1},
-      {9, 7, 4, 0},
-  };
-  std::vector<int> ranks_ = {10, 11, 12, 13, 14, 15, 16, 17, 18, 19};
-  std::vector<std::set<int>> sorted_ranks = {
-      {13},
-      {16, 12},
-      {18, 15, 11},
-      {19, 17, 14, 10},
-  };
-  BOOST_TEST(results_.size() == n);
-  BOOST_TEST(ranks_.size() == n);
-
-  Kokkos::View<int *, DeviceType> ids("query_ids", n);
-  auto ids_host = Kokkos::create_mirror_view(ids);
-  for (int i = 0; i < n; ++i)
-    ids_host(i) = ids_[i];
-  Kokkos::deep_copy(ids, ids_host);
-
-  Kokkos::View<int *, DeviceType> results("results", n);
-  auto results_host = Kokkos::create_mirror_view(results);
-  for (int i = 0; i < n; ++i)
-    results_host(i) = results_[i];
-  Kokkos::deep_copy(results, results_host);
-
-  Kokkos::View<int *, DeviceType> ranks("ranks", n);
-  auto ranks_host = Kokkos::create_mirror_view(ranks);
-  for (int i = 0; i < n; ++i)
-    ranks_host(i) = ranks_[i];
-  Kokkos::deep_copy(ranks, ranks_host);
-
-  ArborX::Details::DistributedSearchTreeImpl<DeviceType>::sortResults(
-      ids, results, ranks);
-
-  // COMMENT: ids are untouched
-  Kokkos::deep_copy(ids_host, ids);
-  BOOST_TEST(ids_host == ids_, tt::per_element());
-
-  Kokkos::deep_copy(results_host, results);
-  Kokkos::deep_copy(ranks_host, ranks);
-  for (int q = 0; q < m; ++q)
-    for (int i = offset[q]; i < offset[q + 1]; ++i)
-    {
-      BOOST_TEST(sorted_results[q].count(results_host[i]) == 1);
-      BOOST_TEST(sorted_ranks[q].count(ranks_host[i]) == 1);
-    }
-
-  Kokkos::View<int *, DeviceType> not_sized_properly("", m);
-  BOOST_CHECK_THROW(
-      ArborX::Details::DistributedSearchTreeImpl<DeviceType>::sortResults(
-          ids, not_sized_properly),
-      ArborX::SearchException);
-}
-
-BOOST_AUTO_TEST_CASE_TEMPLATE(sort_results_2d, DeviceType, ARBORX_DEVICE_TYPES)
-{
-  std::array<int, 5> ids{4, 2, 1, 3, 5};
-  Kokkos::View<int *, DeviceType> ids_device("ids", 5);
-  auto ids_host = Kokkos::create_mirror_view(ids_device);
-  for (int i = 0; i < 5; ++i)
-    ids_host[i] = ids[i];
-  Kokkos::deep_copy(ids_device, ids_host);
-
-  int results_2d[5][2] = {{1, 2}, {3, 4}, {5, 6}, {7, 8}, {9, 10}};
-  int sorted_results_2d[5][2] = {{5, 6}, {3, 4}, {7, 8}, {1, 2}, {9, 10}};
-  Kokkos::View<int **, DeviceType> results_2d_device("results", 5, 2);
-  auto results_2d_host = Kokkos::create_mirror_view(results_2d_device);
-  for (int i = 0; i < 5; ++i)
-    for (int j = 0; j < 2; ++j)
-      results_2d_host(i, j) = results_2d[i][j];
-  Kokkos::deep_copy(results_2d_device, results_2d_host);
-
-  ArborX::Details::DistributedSearchTreeImpl<DeviceType>::sortResults(
-      ids_device, results_2d_device);
-
-  Kokkos::deep_copy(results_2d_host, results_2d_device);
-  for (int i = 0; i < 5; ++i)
-    for (int j = 0; j < 2; ++j)
-      BOOST_TEST(results_2d_host(i, j) == sorted_results_2d[i][j]);
-}
-
-BOOST_AUTO_TEST_CASE_TEMPLATE(sort_results_3d, DeviceType, ARBORX_DEVICE_TYPES)
-{
-  std::array<int, 5> ids{4, 2, 1, 3, 5};
-  Kokkos::View<int *, DeviceType> ids_device("ids", 5);
-  auto ids_host = Kokkos::create_mirror_view(ids_device);
-  for (int i = 0; i < 5; ++i)
-    ids_host[i] = ids[i];
-  Kokkos::deep_copy(ids_device, ids_host);
-
-  int results_3d[5][2][2] = {{{1, 2}, {3, 4}},
-                             {{5, 6}, {7, 8}},
-                             {{9, 10}, {11, 12}},
-                             {{13, 14}, {15, 16}},
-                             {{17, 18}, {19, 20}}};
-  int sorted_results_3d[5][2][2] = {{{9, 10}, {11, 12}},
-                                    {{5, 6}, {7, 8}},
-                                    {{13, 14}, {15, 16}},
-                                    {{1, 2}, {3, 4}},
-                                    {{17, 18}, {19, 20}}};
-  Kokkos::View<int ***, DeviceType> results_3d_device("results", 5, 2, 2);
-  auto results_3d_host = Kokkos::create_mirror_view(results_3d_device);
-  for (int i = 0; i < 5; ++i)
-    for (int j = 0; j < 2; ++j)
-      for (int k = 0; k < 2; ++k)
-        results_3d_host(i, j, k) = results_3d[i][j][k];
-  Kokkos::deep_copy(results_3d_device, results_3d_host);
-
-  ArborX::Details::DistributedSearchTreeImpl<DeviceType>::sortResults(
-      ids_device, results_3d_device);
-
-  Kokkos::deep_copy(results_3d_host, results_3d_device);
-  for (int i = 0; i < 5; ++i)
-    for (int j = 0; j < 2; ++j)
-      for (int k = 0; k < 2; ++k)
-        BOOST_TEST(results_3d_host(i, j, k) == sorted_results_3d[i][j][k]);
-}
 
 BOOST_AUTO_TEST_CASE_TEMPLATE(count_results, DeviceType, ARBORX_DEVICE_TYPES)
 {
